separate out test_head
